Move Fhp\DataTypes\Bin to Fhp\Syntax\Bin
All data types (also dates, for instance) are part of the syntax specification. Since we only have one explicitly implemented, it's not worth having a separate namespace only for the Bin class. Note that things like Kik/Kti/Ktv are *not* syntax-level data types, but just regular DEGs and thus defined in Fhp\Segment\Common.
