+++ lib/Fhp/Segment/HNVSD.php
@@ -2,7 +2,7 @@
 
 namespace Fhp\Segment;
 
-use Fhp\DataTypes\Bin;
+use Fhp\Syntax\Bin;
 
 /**
  * Class HNVSD (Verschlüsselte Daten)
